fix(desktop): use a main entry point and modal response indices in the swift shell
Top-level code cannot call a main-actor initializer, and NSAlert reports button positions relative to alertFirstButtonReturn.
