--- turbo/apps/desktop/native/OkouDesktop/Sources/OkouDesktopApp/main.swift
@@ -1,13 +1,0 @@
-#if canImport(AppKit)
-import AppKit
-
-let delegate = OkouAppDelegate()
-let application = NSApplication.shared
-application.delegate = delegate
-application.run()
-#else
-import Foundation
-
-FileHandle.standardError.write(Data("okou-desktop only runs on macOS.\n".utf8))
-exit(1)
-#endif
